DEV: Move hooks to own folders
