+++ src/hooks/useDictionary/index.js
@@ -1,8 +1,9 @@
 // @flow
 import { useContext } from "react";
 
-import DictionaryContext from "../Dictionary/DictionaryContext";
-import type { UseDictionary } from "./useDictionary";
+import DictionaryContext from "../../Dictionary/DictionaryContext";
+
+import type { UseDictionary } from ".";
 
 const useDictionary: UseDictionary = () => useContext(DictionaryContext);
