+++ src/hooks/useDictionary/index.js.flow
@@ -1,6 +1,6 @@
 // @flow
 
-import type { Translations } from "../Dictionary";
+import type { Translations } from "../../Dictionary";
 
 export type UseDictionary = () => Translations;
