+++ src/hooks/useTheme/index.js
@@ -2,8 +2,9 @@
 import { useContext } from "react";
 import { ThemeContext } from "styled-components";
 
-import defaultTheme from "../defaultTheme";
-import type { UseTheme } from "./useTheme";
+import defaultTheme from "../../defaultTheme";
+
+import type { UseTheme } from ".";
 
 const useTheme: UseTheme = () => useContext(ThemeContext) || defaultTheme;
