+++ src/hooks/useTheme/index.js.flow
@@ -1,5 +1,5 @@
 // @flow
-import type { Theme } from "../defaultTheme";
+import type { Theme } from "../../defaultTheme";
 
 export type UseTheme = () => Theme;
